Extract hit test logic for BarChart out of Painter file for better code readability.
