+++ lib/src/charts/painters/cartesian/bar_painter.dart
@@ -2,15 +2,15 @@ import 'package:chart_it/src/charts/constants/defaults.dart';
 import 'package:chart_it/src/charts/data/bars.dart';
 import 'package:chart_it/src/charts/data/core.dart';
 import 'package:chart_it/src/charts/data/core/cartesian/cartesian_data_internal.dart';
+import 'package:chart_it/src/charts/interactors/cartesian/bar_hit_tester.dart';
 import 'package:chart_it/src/charts/painters/cartesian/cartesian_painter.dart';
 import 'package:chart_it/src/charts/painters/text/chart_text_painter.dart';
-import 'package:chart_it/src/extensions/interactions.dart';
 import 'package:chart_it/src/extensions/paint_objects.dart';
 import 'package:chart_it/src/interactions/interactions.dart';
 import 'package:flutter/material.dart';
 
 class BarPainter implements CartesianPainter<BarInteractionResult> {
-  final List<_BarGroupInteractionWrapper> _groupInteractions =
+  final List<BarGroupInteractionData> _groupInteractions =
       List.empty(growable: true);
 
   _BarPainterData? _data;
@@ -38,233 +38,16 @@ class BarPainter implements CartesianPainter<BarInteractionResult> {
       var snapToBarConfig = data.series.interactionEvents.snapToBarConfig;
       var fuzziness = data.series.interactionEvents.fuzziness;
 
-      _BarGroupInteractionWrapper? previousGroup;
-
-      for (var i = 0; i < _groupInteractions.length; i++) {
-        final isLastGroup = i == _groupInteractions.length - 1;
-        final group = _groupInteractions[i];
-
-        if (group.isInteractionWithinBounds(localPosition)) {
-          return _getBarInteraction(
-            group: group,
-            snapToBarConfig: snapToBarConfig,
-            type: type,
-            localPosition: localPosition,
-            fuzziness: fuzziness,
-            groupType: group.type,
-          );
-        }
-
-        if (!group.isPointerAfterCurrentGroup(localPosition) || isLastGroup) {
-          return _snapAndGetBarInteraction(
-            group: group,
-            previousGroup: previousGroup,
-            snapToBarConfig: snapToBarConfig,
-            type: type,
-            localPosition: localPosition,
-            fuzziness: fuzziness,
-            graphUnitWidth: data.graphUnitWidth,
-          );
-        }
-
-        previousGroup = group;
-      }
+      return BarHitTester.hitTest(
+        graphUnitWidth: data.graphUnitWidth,
+        localPosition: localPosition,
+        type: type,
+        interactionData: _groupInteractions,
+        snapToBarConfig: snapToBarConfig,
+        fuzziness: fuzziness,
+      );
     }
     // No Interactions for this BarSeries.
-    return null;
-  }
-
-  BarInteractionResult? _snapAndGetBarInteraction({
-    required _BarGroupInteractionWrapper group,
-    required _BarGroupInteractionWrapper? previousGroup,
-    required SnapToBarConfig snapToBarConfig,
-    required TouchInteractionType type,
-    required Offset localPosition,
-    required Fuzziness fuzziness,
-    required double graphUnitWidth,
-  }) {
-    final _BarGroupInteractionWrapper interactionGroup;
-
-    switch (snapToBarConfig.snapToBarBehaviour) {
-      case SnapToBarBehaviour.snapToNearest:
-        interactionGroup = _findNearestGroupByDistance(
-          localPosition: localPosition,
-          previousGroup: previousGroup,
-          currentGroup: group,
-        );
-        break;
-      case SnapToBarBehaviour.snapToSection:
-        interactionGroup = _findNearestGroupBySection(
-          localPosition: localPosition,
-          previousGroup: previousGroup,
-          currentGroup: group,
-          graphUnitWidth: graphUnitWidth,
-        );
-        break;
-    }
-
-    return _getBarInteraction(
-      group: interactionGroup,
-      snapToBarConfig: snapToBarConfig,
-      type: type,
-      localPosition: localPosition,
-      fuzziness: fuzziness,
-      groupType: interactionGroup.type,
-    );
-  }
-
-  BarInteractionResult? _getBarInteraction(
-      {required _BarGroupInteractionWrapper group,
-      required SnapToBarConfig snapToBarConfig,
-      required TouchInteractionType type,
-      required Offset localPosition,
-      required Fuzziness fuzziness,
-      required GroupType groupType}) {
-    switch (groupType) {
-      case GroupType.simpleBar:
-        return _getSeriesBarInteraction(
-          group: group,
-          snapToBarConfig: snapToBarConfig,
-          type: type,
-          localPosition: localPosition,
-          fuzziness: fuzziness,
-        );
-      case GroupType.multiBarSeries:
-        return _getSeriesBarInteraction(
-          group: group,
-          snapToBarConfig: snapToBarConfig,
-          type: type,
-          localPosition: localPosition,
-          fuzziness: fuzziness,
-        );
-      case GroupType.multiBarStack:
-        return _getStackBarInteractionResult(
-          group: group,
-          snapToBarConfig: snapToBarConfig,
-          type: type,
-          localPosition: localPosition,
-          fuzziness: fuzziness,
-        );
-    }
-  }
-
-  _BarGroupInteractionWrapper _findNearestGroupByDistance({
-    required Offset localPosition,
-    required _BarGroupInteractionWrapper? previousGroup,
-    required _BarGroupInteractionWrapper currentGroup,
-  }) {
-    final distToPrev =
-        (localPosition.dx - (previousGroup?.groupEnd ?? double.maxFinite))
-            .abs();
-    final distToNext = (currentGroup.groupStart - localPosition.dx).abs();
-
-    if (distToPrev < distToNext && previousGroup != null) {
-      return previousGroup;
-    } else {
-      return currentGroup;
-    }
-  }
-
-  _BarGroupInteractionWrapper _findNearestGroupBySection({
-    required Offset localPosition,
-    required _BarGroupInteractionWrapper? previousGroup,
-    required _BarGroupInteractionWrapper currentGroup,
-    required double graphUnitWidth,
-  }) {
-    final index = previousGroup?.groupIndex ?? currentGroup.groupIndex;
-    final widthMultiplicationFactor = index + 1;
-    final currentUnitWidthEndOffset =
-        graphUnitWidth * widthMultiplicationFactor;
-
-    final _BarGroupInteractionWrapper interactionGroup;
-    if (localPosition.dx < currentUnitWidthEndOffset) {
-      interactionGroup = previousGroup ?? currentGroup;
-    } else {
-      interactionGroup = currentGroup;
-    }
-
-    return interactionGroup;
-  }
-
-  BarInteractionResult? _getSeriesBarInteraction({
-    required _BarGroupInteractionWrapper group,
-    required SnapToBarConfig snapToBarConfig,
-    required TouchInteractionType type,
-    required Offset localPosition,
-    required Fuzziness fuzziness,
-  }) {
-    final bars = group.barInteractions;
-    _BarInteractionData? previousBar;
-    for (var i = 0; i < bars.length; i++) {
-      final bar = bars[i];
-      final shouldSnapToHeight = snapToBarConfig.shouldSnapToHeight(type);
-
-      if (bar.containsWithFuzziness(
-        position: localPosition,
-        fuzziness: fuzziness,
-        snapToHeight: shouldSnapToHeight,
-      )) {
-        return bar.getInteractionResult(localPosition, type);
-      }
-
-      if (snapToBarConfig.shouldSnapToWidth(type)) {
-        final isLastBar = i == bars.length - 1;
-        final isPointerAfterBar = bar.isPointerAfterBar(localPosition);
-        // The condition here is
-        // 1. If Pointer is after bar and
-        // 2. Is not the last bar.
-        if (!isPointerAfterBar || isLastBar) {
-          return _snapToNearestBar(
-            localPosition: localPosition,
-            fuzziness: fuzziness,
-            type: type,
-            previousBar: previousBar,
-            currentBar: bar,
-            isPointerAfterBar: isPointerAfterBar,
-            isLastBar: isLastBar,
-            shouldSnapToHeight: shouldSnapToHeight,
-            groupType: group.type,
-          );
-        } else {
-          previousBar = bar;
-        }
-      }
-    }
-
-    return null;
-  }
-
-  BarInteractionResult? _getStackBarInteractionResult({
-    required _BarGroupInteractionWrapper group,
-    required SnapToBarConfig snapToBarConfig,
-    required TouchInteractionType type,
-    required Offset localPosition,
-    required Fuzziness fuzziness,
-  }) {
-    final bars = group.barInteractions;
-    for (var i = 0; i < bars.length; i++) {
-      final bar = bars[i];
-      final isLastBar = i == bars.length - 1;
-      final shouldSnapToHeight =
-          snapToBarConfig.shouldSnapToHeight(type) && isLastBar;
-
-      if (bar.containsWithFuzziness(
-        position: localPosition,
-        fuzziness: fuzziness,
-        snapToHeight: shouldSnapToHeight,
-      )) {
-        return bar.getInteractionResult(localPosition, type);
-      }
-
-      if (snapToBarConfig.shouldSnapToWidth(type)) {
-        final isWithinBarHeight = bar.rect.top <= localPosition.dy &&
-            localPosition.dy <= bar.rect.bottom;
-        if (isWithinBarHeight || (isLastBar && shouldSnapToHeight)) {
-          return bar.getInteractionResult(localPosition, type);
-        }
-      }
-    }
-
     return null;
   }
 
@@ -360,7 +143,7 @@ class BarPainter implements CartesianPainter<BarInteractionResult> {
         data.series.seriesStyle ??
         defaultBarSeriesStyle;
 
-    final barInteractions = <_BarInteractionData>[];
+    final barInteractions = <BarInteractionData>[];
     final dxPos = dxOffset + (data.unitWidth * 0.5) - (data.barWidth * 0.5);
 
     // Since we have only one yValue, we only have to draw one bar
@@ -394,10 +177,10 @@ class BarPainter implements CartesianPainter<BarInteractionResult> {
       dx: dxOffset,
     );
     // Before we move on to the next group, we have to save the interaction data for this group
-    final groupInteraction = _BarGroupInteractionWrapper(
+    final groupInteraction = BarGroupInteractionData(
       groupStart: dxPos,
       groupEnd: dxPos + data.barWidth,
-      type: GroupType.simpleBar,
+      type: InteractedGroupType.simpleBar,
       groupIndex: groupIndex,
       barInteractions: barInteractions,
     );
@@ -413,7 +196,7 @@ class BarPainter implements CartesianPainter<BarInteractionResult> {
     required CartesianChartStylingData style,
     required _BarPainterData data,
   }) {
-    final barInteractions = <_BarInteractionData>[];
+    final barInteractions = <BarInteractionData>[];
     // var groupWidth = _unitWidth / group.yValues.length;
     // Draw individual bars in this group
     var groupCount = group.yValues.length; // No. of groups for this multibar
@@ -467,11 +250,10 @@ class BarPainter implements CartesianPainter<BarInteractionResult> {
       x += data.barWidth;
     }
     // Before we move on to the next group, we have to save the interaction data for this group
-    final groupInteraction = _BarGroupInteractionWrapper(
+    final groupInteraction = BarGroupInteractionData(
       groupStart: groupStart,
       groupEnd: x,
-      // dxOffset + data.unitWidth,
-      type: GroupType.multiBarSeries,
+      type: InteractedGroupType.multiBarSeries,
       groupIndex: groupIndex,
       barInteractions: barInteractions,
     );
@@ -487,7 +269,7 @@ class BarPainter implements CartesianPainter<BarInteractionResult> {
     required CartesianChartStylingData style,
     required _BarPainterData data,
   }) {
-    final barInteractions = <_BarInteractionData>[];
+    final barInteractions = <BarInteractionData>[];
     var stackCount = group.yValues.length; // No. of bars for this stack
 
     final dxPos = dxOffset + (data.unitWidth * 0.5) - (data.barWidth * 0.5);
@@ -543,10 +325,10 @@ class BarPainter implements CartesianPainter<BarInteractionResult> {
       }
     }
     // Before we move on to the next group, we have to save the interaction data for this group
-    final groupInteraction = _BarGroupInteractionWrapper(
+    final groupInteraction = BarGroupInteractionData(
       groupStart: dxPos,
       groupEnd: dxPos + data.barWidth,
-      type: GroupType.multiBarStack,
+      type: InteractedGroupType.multiBarStack,
       groupIndex: groupIndex,
       barInteractions: barInteractions,
     );
@@ -567,7 +349,7 @@ class BarPainter implements CartesianPainter<BarInteractionResult> {
     required double leftPadding,
     required double rightPadding,
     required _BarPainterData data,
-    required Function(_BarInteractionData) onShapeDimensionsReady,
+    required Function(BarInteractionData) onShapeDimensionsReady,
   }) {
     // The first thing to do is to get the data point into the range!
     // This is because we don't want our bar to exceed the min/max values
@@ -598,7 +380,7 @@ class BarPainter implements CartesianPainter<BarInteractionResult> {
     );
 
     // Before we Paint this Bar, we need to store it's data for hitTesting
-    var shapeData = _BarInteractionData(
+    var shapeData = BarInteractionData(
       rect: bar,
       barGroup: barGroup,
       barGroupIndex: barGroupIndex,
@@ -677,212 +459,6 @@ class BarPainter implements CartesianPainter<BarInteractionResult> {
       );
     }
   }
-
-  BarInteractionResult? _snapToNearestBar({
-    required Offset localPosition,
-    required Fuzziness fuzziness,
-    required TouchInteractionType type,
-    required _BarInteractionData? previousBar,
-    required _BarInteractionData currentBar,
-    required bool isPointerAfterBar,
-    required bool isLastBar,
-    required bool shouldSnapToHeight,
-    required GroupType groupType,
-  }) {
-    final isFirstBar = previousBar == null;
-    final isPointerAfterLastBar = isLastBar && isPointerAfterBar;
-
-    if (isFirstBar || isPointerAfterLastBar) {
-      if (_ignorePointerAboveBar(
-        shouldSnapToHeight: shouldSnapToHeight,
-        bar: currentBar,
-        localPosition: localPosition,
-        fuzziness: fuzziness,
-      )) {
-        return null;
-      }
-
-      return currentBar.getInteractionResult(localPosition, type);
-    }
-
-    final _BarInteractionData nearestBar;
-
-    switch (groupType) {
-      case GroupType.simpleBar:
-        nearestBar =
-            _findNearestBarWithDx(localPosition, previousBar, currentBar);
-        break;
-      case GroupType.multiBarSeries:
-        nearestBar =
-            _findNearestBarWithDx(localPosition, previousBar, currentBar);
-        break;
-      case GroupType.multiBarStack:
-        nearestBar =
-            _findNearestBarWithDy(localPosition, previousBar, currentBar);
-        break;
-    }
-
-    // If snap to height is disabled and pointer is above the nearest bar, return null
-    if (_ignorePointerAboveBar(
-      shouldSnapToHeight: shouldSnapToHeight,
-      bar: nearestBar,
-      localPosition: localPosition,
-      fuzziness: fuzziness,
-    )) {
-      return null;
-    }
-
-    return nearestBar.getInteractionResult(localPosition, type);
-  }
-
-  bool _ignorePointerAboveBar({
-    required bool shouldSnapToHeight,
-    required _BarInteractionData bar,
-    required Offset localPosition,
-    required Fuzziness fuzziness,
-  }) {
-    return !shouldSnapToHeight &&
-        bar.isPointerAboveBar(
-          position: localPosition,
-          fuzziness: fuzziness,
-        );
-  }
-
-  // BarInteractionResult? _snapToSectionBar({
-  //   required _BarPainterData data,
-  //   required Offset localPosition,
-  //   required _BarInteractionData? previousBar,
-  //   required _BarInteractionData bar,
-  //   required TouchInteractionType type,
-  //   required Fuzziness fuzziness,
-  //   required bool isLastBar,
-  //   required bool shouldSnapToHeight,
-  // }) {
-  //   final index = previousBar?.barGroupIndex ?? bar.barGroupIndex;
-  //   final widthMultiplicationFactor = index + 1;
-  //   final currentUnitWidthEndOffset =
-  //       data.graphUnitWidth * widthMultiplicationFactor;
-  //
-  //   if (localPosition.dx <= currentUnitWidthEndOffset) {
-  //     if (previousBar?.barGroupIndex == bar.barGroupIndex) {
-  //       return _snapToNearestBar(
-  //         localPosition: localPosition,
-  //         fuzziness: fuzziness,
-  //         type: type,
-  //         previousBar: previousBar,
-  //         currentBar: bar,
-  //         isPointerAfterBar: false,
-  //         isLastBar: isLastBar,
-  //         shouldSnapToHeight: shouldSnapToHeight,
-  //       );
-  //     }
-  //
-  //     return previousBar?.getInteractionResult(localPosition, type) ??
-  //         bar.getInteractionResult(localPosition, type);
-  //   } else {
-  //     return bar.getInteractionResult(localPosition, type);
-  //   }
-  // }
-
-  _BarInteractionData _findNearestBarWithDx(
-    Offset position,
-    _BarInteractionData previous,
-    _BarInteractionData current,
-  ) {
-    final distToPrev = position.dx - previous.rect.right;
-    final distToCurrent = current.rect.left - position.dx;
-    return distToPrev >= distToCurrent ? current : previous;
-  }
-
-  _BarInteractionData _findNearestBarWithDy(
-    Offset position,
-    _BarInteractionData previous,
-    _BarInteractionData current,
-  ) {
-    // We have two scenarios over here
-    // 1.  The pointer is alongside the bar.
-    // 2. The pointer is over the top most bar.
-    final previousHeightBounds = _isWithinHeightBounds(position, previous);
-    final currentHeightBounds = _isWithinHeightBounds(position, current);
-
-    if (previousHeightBounds || currentHeightBounds) {
-      // Pointer is alongside the bar
-      return currentHeightBounds ? current : previous;
-    } else {
-      // Pointer is on the top most bar
-      // in this case, we need to find the closest bar by the nearest top side of the bar
-      final distToPrev = position.dy - previous.rect.top;
-      final distToCurrent = position.dy - current.rect.top;
-      return distToPrev >= distToCurrent ? current : previous;
-    }
-  }
-
-  bool _isWithinHeightBounds(Offset position, _BarInteractionData bar) {
-    return bar.rect.top <= position.dy && bar.rect.bottom >= position.dy;
-  }
-}
-
-class _BarInteractionData {
-  final RRect rect;
-  final BarGroup barGroup;
-  final int barGroupIndex;
-  final BarData barData;
-  final int barDataIndex;
-
-  _BarInteractionData({
-    required this.rect,
-    required this.barGroup,
-    required this.barGroupIndex,
-    required this.barData,
-    required this.barDataIndex,
-  });
-
-  bool containsWithFuzziness({
-    required Offset position,
-    required Fuzziness fuzziness,
-    required bool snapToHeight,
-  }) {
-    final left = rect.left - fuzziness.left;
-    final right = rect.right + fuzziness.right;
-
-    if (snapToHeight) {
-      // If snapToHeight is true, any value of dy should return true. So we just check for dx constraints.
-      return position.dx >= left && position.dx < right;
-    }
-
-    final top = rect.top - fuzziness.top;
-    final bottom = rect.bottom + fuzziness.bottom;
-    return position.dx >= left &&
-        position.dx < right &&
-        position.dy >= top &&
-        position.dy < bottom;
-  }
-
-  bool isPointerAfterBar(Offset position) => position.dx > rect.right;
-
-  bool isPointerAboveBar({
-    required Offset position,
-    required Fuzziness fuzziness,
-  }) {
-    final top = rect.top - fuzziness.top;
-    final bottom = rect.bottom + fuzziness.bottom;
-
-    return position.dy < top || position.dy > bottom;
-  }
-
-  BarInteractionResult getInteractionResult(
-    Offset localPosition,
-    TouchInteractionType type,
-  ) {
-    return BarInteractionResult(
-      barGroup: barGroup,
-      barGroupIndex: barGroupIndex,
-      barData: barData,
-      barDataIndex: barDataIndex,
-      localPosition: localPosition,
-      interactionType: type,
-    );
-  }
 }
 
 class _BarPainterData {
@@ -907,28 +483,4 @@ class _BarPainterData {
     required this.graphUnitWidth,
     required this.valueUnitWidth,
   });
-}
-
-enum GroupType { simpleBar, multiBarSeries, multiBarStack }
-
-class _BarGroupInteractionWrapper {
-  final double groupStart;
-  final double groupEnd;
-  final GroupType type;
-  final int groupIndex;
-  final List<_BarInteractionData> barInteractions;
-
-  _BarGroupInteractionWrapper({
-    required this.groupStart,
-    required this.groupEnd,
-    required this.type,
-    required this.groupIndex,
-    required this.barInteractions,
-  });
-
-  bool isPointerAfterCurrentGroup(Offset position) => position.dx > groupEnd;
-
-  bool isInteractionWithinBounds(Offset localPosition) {
-    return groupStart <= localPosition.dx && groupEnd >= localPosition.dx;
-  }
 }
